# api/server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "MLX API Extension for Kubeflow Pipelines"
  version: "0.1.29-dont-cache-kfservices"
  title: "MLX API"
basePath: "/apis/v1alpha1"
schemes:
- "http"
paths:
...
  /pipelines:
    get:
      tags:
      - "PipelineService"
      operationId: "list_pipelines"
      parameters:
      - name: "page_token"
        in: "query"
        required: false
        type: "string"
      - name: "page_size"
        in: "query"
        required: false
        type: "integer"
        format: "int32"
      - name: "sort_by"
        in: "query"
        description: "Can be format of \"field_name\", \"field_name asc\" or \"field_name\
          \ desc\"\nAscending by default."
        required: false
        type: "string"
      - name: "filter"
        in: "query"
        description: "A string-serialized JSON dictionary with key-value pairs that\
          \ correspond to the Pipeline's attribute names and their respective values\
          \ to be filtered for."
        required: false
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiListPipelinesResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
    post:
      tags:
      - "PipelineService"
      operationId: "create_pipeline"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/apiPipeline"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiPipeline"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/{id}:
    get:
      tags:
      - "PipelineService"
      operationId: "get_pipeline"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiPipelineExtended"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
    delete:
      tags:
      - "PipelineService"
      operationId: "delete_pipeline"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "204":
          description: "Deletion successful."
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/{id}/download:
    get:
      tags:
      - "PipelineService"
      summary: "Returns the pipeline YAML compressed into a .tgz (.tar.gz) file."
      operationId: "download_pipeline_files"
      produces:
      - "application/gzip"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: "A TGZ file containing the component artifacts."
          schema:
            type: "file"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/{id}/run:
    post:
      tags:
      - "PipelineService"
      operationId: "run_pipeline"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "run_name"
        in: "query"
        description: "name to identify the run on the Kubeflow Pipelines UI, defaults\
          \ to pipeline name"
        required: false
        type: "string"
      - in: "body"
        name: "parameters"
        description: "optional run parameters, may be required based on pipeline definition"
        required: false
        schema:
          $ref: "#/definitions/dictionary"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiRunCodeResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/{id}/templates:
    get:
      tags:
      - "PipelineService"
      operationId: "get_template"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiGetTemplateResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/upload:
    post:
      tags:
      - "PipelineService"
      operationId: "upload_pipeline"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "uploadfile"
        in: "formData"
        description: "The pipeline YAML file to upload. Can be a GZip-compressed TAR\
          \ file (.tgz, .tar.gz) or a YAML file (.yaml, .yml). Maximum size is 32MB."
        required: true
        type: "file"
      - name: "name"
        in: "query"
        description: "A name for this pipeline, optional"
        required: false
        type: "string"
      - name: "description"
        in: "query"
        description: "A description for this pipeline, optional"
        required: false
        type: "string"
      - name: "annotations"
        in: "formData"
        description: "A string representation of a JSON dictionary of annotations\
          \ describing this pipeline, optional. Example: {\"platform\": \"Kubeflow\"\
          , \"license\": \"Opensource\"}"
        required: false
        type: "string"
      responses:
        "201":
          description: ""
          schema:
            $ref: "#/definitions/apiPipelineExtended"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/upload_from_url:
    post:
      tags:
      - "PipelineService"
      operationId: "upload_pipeline_from_url"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "url"
        in: "formData"
        description: "URL pointing to the pipeline YAML file."
        required: true
        type: "string"
      - name: "name"
        in: "query"
        description: "Optional, the name of the pipeline to be created overriding\
          \ the name in the YAML file."
        required: false
        type: "string"
      - name: "access_token"
        in: "formData"
        description: "Optional, the Bearer token to access the 'url'."
        required: false
        type: "string"
      responses:
        "201":
          description: "Pipeline was created successfully."
          schema:
            $ref: "#/definitions/apiPipeline"
        default:
          description: "Something went wrong."
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/featured:
    post:
      tags:
      - "PipelineService"
      operationId: "set_featured_pipelines"
      parameters:
      - in: "body"
        name: "pipeline_ids"
        description: "Array of pipeline IDs to be featured."
        required: true
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        "200":
          description: "Success"
        default:
          description: "Something went wrong"
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/publish_approved:
    post:
      tags:
      - "PipelineService"
      operationId: "approve_pipelines_for_publishing"
      parameters:
      - in: "body"
        name: "pipeline_ids"
        description: "Array of pipeline IDs to be approved for publishing."
        required: true
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        "200":
          description: "Success"
        default:
          description: "Something went wrong"
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
  /pipelines/run_custom_pipeline:
    post:
      tags:
      - "PipelineService"
      description: "Run a complex pipeline defined by a directed acyclic graph (DAG)"
      operationId: "run_custom_pipeline"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "run_custom_pipeline_payload"
        description: "A custom pipeline defined by a directed acyclic graph (DAG)\
          \ and input parameters"
        required: true
        schema:
          $ref: "#/definitions/apiPipelineCustomRunPayload"
      - name: "run_name"
        in: "query"
        description: "Name to identify the run on the Kubeflow Pipelines UI"
        required: false
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiRunCodeResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.pipeline_service_controller"
definitions:
...
  apiGetTemplateResponse:
    type: "object"
    properties:
      template:
        type: "string"
        description: "The YAML template file content"
      url:
        type: "string"
        description: "The URL to download the template text from S3 storage (Minio)"
    example:
      template: "TODO: YAML file content"
      url: "http://min.io:9000/mlpipeline/components/1lBnSAce8lAXP97Yb11eqSk/train_wml.yaml"
...
  apiListPipelinesResponse:
    type: "object"
    properties:
      pipelines:
        type: "array"
        items:
          $ref: "#/definitions/apiPipelineExtended"
      total_size:
        type: "integer"
        format: "int32"
      next_page_token:
        type: "string"
    example:
      pipelines:
      - name: "name"
        created_at: "2018-12-29T04:56:07.000+00:00"
        description: "description"
        id: "e3U42y3-au240nc-13nc"
        status: "READY"
        parameters:
        - name: "name"
          value: "value"
        - name: "name"
          value: "value"
      - name: "another pipeline"
        created_at: "2019-06-23T14:18:01.000+00:00"
        description: "another pipeline"
        id: "v92-834nat-2ndida"
        annotations:
          platform: "KFP"
      next_page_token: "next_page_token"
      total_size: 10
...
  apiParameter:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      description:
        type: "string"
      default:
        $ref: "#/definitions/anyValue"
      value:
        $ref: "#/definitions/anyValue"
    example:
      name: "name"
      value: "value"
  apiPipeline:
    type: "object"
    properties:
      id:
        type: "string"
      created_at:
        type: "string"
        format: "date-time"
      name:
        type: "string"
      description:
        type: "string"
      parameters:
        type: "array"
        items:
          $ref: "#/definitions/apiParameter"
      status:
        type: "string"
        description: "In case any error happens retrieving a pipeline field, only\
          \ pipeline ID\nand the error message is returned. Client has the flexibility\
          \ of choosing\nhow to handle error. This is especially useful during listing\
          \ call."
      default_version_id:
        type: "string"
        description: "The default version of the pipeline. As of now, the latest version\
          \ is used as default. (In the future, if desired by customers, we can allow\
          \ them to set default version.)"
        readOnly: true
      namespace:
        type: "string"
    example:
      name: "name"
      created_at: "2000-01-23T04:56:07.000+00:00"
      description: "description"
      id: "id"
      status: "READY"
      parameters:
      - name: "name"
        value: "value"
      - name: "name"
        value: "value"
  apiPipelineExtension:
    type: "object"
    properties:
      id:
        type: "string"
      annotations:
        type: "object"
        additionalProperties:
          type: "string"
      featured:
        type: "boolean"
      publish_approved:
        type: "boolean"
    example:
      id: "id"
      annotations:
        platform: "KFP"
  apiPipelineExtended:
    allOf:
    - $ref: "#/definitions/apiPipeline"
    - $ref: "#/definitions/apiPipelineExtension"
    description: "A representation of a pipeline with extended properties."
  apiPipelineCustomRunPayload:
    type: "object"
    properties:
      custom_pipeline:
        $ref: "#/definitions/apiPipelineCustom"
      run_parameters:
        $ref: "#/definitions/dictionary"
    description: "Wrapper object to post a custom pipeline run defined by a directed\
      \ acyclic graph (DAG) and input parameters."
  apiPipelineCustom:
    type: "object"
    required:
    - "dag"
    - "name"
    properties:
      dag:
        $ref: "#/definitions/apiPipelineDAG"
      inputs:
        $ref: "#/definitions/apiPipelineInputs"
      name:
        type: "string"
        description: "Name of the custom pipeline"
      description:
        type: "string"
        description: "Optional description of the custom pipeline"
    description: "A custom pipeline defined by a directed acyclic graph (DAG) and\
      \ input parameters."
  apiPipelineInputs:
    type: "object"
    properties:
      parameters:
        type: "array"
        description: "List of pipeline parameters"
        items:
          $ref: "#/definitions/apiParameter"
    description: "Input parameters to custom pipeline"
  apiPipelineDAG:
    type: "object"
    properties:
      tasks:
        type: "array"
        description: "List of pipeline tasks."
        items:
          $ref: "#/definitions/apiPipelineTask"
    description: "A directed acyclic graph (DAG) describing a complex pipeline"
  apiPipelineTask:
    type: "object"
    required:
    - "artifact_id"
    - "artifact_type"
    properties:
      name:
        type: "string"
      artifact_type:
        type: "string"
        description: "The type of artifact for this task, can be either one of: 'component',\
          \ 'model', 'notebook', 'pipeline'"
      artifact_id:
        type: "string"
        description: "The UUID of the artifact for this task"
      arguments:
        $ref: "#/definitions/apiPipelineTaskArguments"
      dependencies:
        type: "array"
        description: "Task dependencies, referring to upstream tasks that have to\
          \ be completed prior to running this task by their respective task names"
        items:
          type: "string"
    description: "A custom pipeline task as part of a directed acyclic graph (DAG)."
  apiPipelineTaskArguments:
    type: "object"
    properties:
      parameters:
        type: "array"
        items:
          $ref: "#/definitions/apiParameter"
    description: "Arguments to a custom pipeline task as part of a directed acyclic\
      \ graph (DAG)."
  apiRunCodeResponse:
    type: "object"
    required:
    - "run_url"
    properties:
      run_url:
        type: "string"
        description: "The URL path to view the run details in the Kubeflow Pipeline\
          \ Web UI"
      run_output_location:
        type: "string"
        description: "The location of the output artifacts produced by the run, if\
          \ any."
    example:
      run_url: "/runs/details/0e2680ec-5d5a-11e9-9911-5ed556129669"
...
  apiStatus:
    type: "object"
    properties:
      error:
        type: "string"
      code:
        type: "integer"
        format: "int32"
      details:
        type: "array"
        items:
          $ref: "#/definitions/protobufAny"
...
  anyValue:
    description: "A non-null value of any type like string, number, boolean, array,\
      \ object, see https://stackoverflow.com/a/43328994/5601796 (swagger-2-0-what-schema-to-accept-any-complex-json-value)"
  dictionary:
    type: "object"
    description: "A dictionary of key-value pairs with values of any type."
    example:
      some_key: "some value"
      another_key: true
      some_number: 42
    additionalProperties:
      $ref: "#/definitions/anyValue"
  protobufAny:
    type: "object"
    properties:
      type_url:
        type: "string"
        description: "TODO"
      value:
        type: "string"
        format: "byte"
        description: "Must be a valid serialized protocol buffer of the above specified\
          \ type."
        pattern: "^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$"
        x-regex: "^(?:[A-Za-z0-9+\\/]{4})*(?:[A-Za-z0-9+\\/]{2}==|[A-Za-z0-9+\\/]{3}=)?$"
        x-modifiers: []
    description: "TODO"
...